backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.24.4
scipy==1.11.4
opencv-python-headless==4.9.0.80
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
